sandbox - move native environment service to common layer
