+++ src/main.js
@@ -200,9 +200,9 @@ async function onReady() {
 }
 
 /**
- * @typedef	 {{ [arg: string]: any; '--'?: string[]; _: string[]; }} ParsedArgs
+ * @typedef	 {{ [arg: string]: any; '--'?: string[]; _: string[]; }} NativeParsedArgs
  *
- * @param {ParsedArgs} cliArgs
+ * @param {NativeParsedArgs} cliArgs
  */
 function configureCommandlineSwitchesSync(cliArgs) {
 	const SUPPORTED_ELECTRON_SWITCHES = [
@@ -355,7 +355,7 @@ function getArgvConfigPath() {
 }
 
 /**
- * @param {ParsedArgs} cliArgs
+ * @param {NativeParsedArgs} cliArgs
  * @returns {string}
  */
 function getJSFlags(cliArgs) {
@@ -375,7 +375,7 @@ function getJSFlags(cliArgs) {
 }
 
 /**
- * @param {ParsedArgs} cliArgs
+ * @param {NativeParsedArgs} cliArgs
  *
  * @returns {string}
  */
@@ -388,7 +388,7 @@ function getUserDataPath(cliArgs) {
 }
 
 /**
- * @returns {ParsedArgs}
+ * @returns {NativeParsedArgs}
  */
 function parseCLIArgs() {
 	const minimist = require('minimist');
