+++ src/vs/platform/extensionManagement/node/extensionLifecycle.ts
@@ -12,7 +12,6 @@ import { join } from 'vs/base/common/path';
 import { Limiter } from 'vs/base/common/async';
 import { Event } from 'vs/base/common/event';
 import { Schemas } from 'vs/base/common/network';
-import { INativeEnvironmentService } from 'vs/platform/environment/node/environmentService';
 import { rimraf } from 'vs/base/node/pfs';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 
@@ -21,7 +20,7 @@ export class ExtensionsLifecycle extends Disposable {
 	private processesLimiter: Limiter<void> = new Limiter(5); // Run max 5 processes in parallel
 
 	constructor(
-		@IEnvironmentService private environmentService: INativeEnvironmentService,
+		@IEnvironmentService private environmentService: IEnvironmentService,
 		@ILogService private readonly logService: ILogService
 	) {
 		super();
